refactor: 기존 useProduct 복구 및 나머지 훅 임시 구현
